requirements.txt
django
djangorestframework
jsonfield
# django-cors-headers